add code coverage and refactor tests folder
