--- mocha.opts
@@ -1,1 +1,0 @@
---compilers js:babel-core/register
