+++ src/store/history.js
@@ -1,3 +1,13 @@
 import createHashHistory from 'history/lib/createHashHistory';
 
-export default createHashHistory();
+const _createHashHistory = () => {
+  if (global.window) {
+    return createHashHistory();
+  }
+
+  return undefined;
+};
+
+const createHistory = _createHashHistory();
+
+export default createHistory;
