+++ tests/mocha.opts
@@ -1,0 +1,3 @@
+--require ./tests/entry.js
+--compilers js:babel-core/register
+tests/**/*.test.js src/**/*.test.js
